update: improve UI and UX of the map editor
